# specs/001-srs-ag-school-finance/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: AG School Finance API Specification
  description: OpenAPI REST contracts for AG School Finance (Public Transparency & Internal Operations)
  version: 1.0.0
servers:
  - url: http://localhost/agschool/api
    description: Local Laragon Server
paths:
...
  /auth/login:
    post:
      summary: User login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
...